feat: Implement new cloud backends

+++ source/core/data/src/main/kotlin/com/xayah/core/data/repository/CloudRepository.kt
@@ -2,17 +2,12 @@ package com.xayah.core.data.repository
 
 import android.content.Context
 import androidx.annotation.StringRes
-import com.google.gson.reflect.TypeToken
 import com.xayah.core.database.dao.CloudDao
-import com.xayah.core.database.model.AccountMap
-import com.xayah.core.database.model.CloudAccountEntity
 import com.xayah.core.database.model.CloudEntity
+import com.xayah.core.network.client.CloudClient
 import com.xayah.core.rootservice.service.RemoteRootService
-import com.xayah.core.util.GsonUtil
 import com.xayah.core.util.LogUtil
 import com.xayah.core.util.PathUtil
-import com.xayah.core.util.command.BaseUtil
-import com.xayah.core.util.command.Rclone
 import com.xayah.core.util.model.ShellResult
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.flow.distinctUntilChanged
@@ -22,8 +17,6 @@ class CloudRepository @Inject constructor(
     @ApplicationContext private val context: Context,
     private val rootService: RemoteRootService,
     private val cloudDao: CloudDao,
-    private val gsonUtil: GsonUtil,
-    private val pathUtil: PathUtil,
 ) {
     private fun log(msg: () -> String): String = run {
         LogUtil.log { "CloudRepository" to msg() }
@@ -31,67 +24,28 @@ class CloudRepository @Inject constructor(
     }
 
     fun getString(@StringRes resId: Int) = context.getString(resId)
-    suspend fun upsertCloud(item: CloudEntity) = cloudDao.upsertCloud(item)
-    suspend fun queryCloudByName(name: String) = cloudDao.queryCloudByName(name)
+    suspend fun upsert(item: CloudEntity) = cloudDao.upsert(item)
+    suspend fun queryByName(name: String) = cloudDao.queryByName(name)
 
-    val clouds = cloudDao.queryActiveCloudsFlow().distinctUntilChanged()
+    val clouds = cloudDao.queryFlow().distinctUntilChanged()
 
-    suspend fun update() {
-        Rclone.Config.dump().also { result ->
-            val type = object : TypeToken<AccountMap>() {}.type
-            val accountMap = runCatching { gsonUtil.fromJson<AccountMap>(result.outString, type) }.onFailure {
-                val msg = it.message
-                if (msg != null)
-                    log { msg }
-            }.getOrElse { AccountMap() }
-            // Inactivate all cloud entities.
-            cloudDao.updateActive(false)
-            cloudDao.upsertAccount(accountMap.map { (key, value) ->
-                CloudAccountEntity(
-                    name = key,
-                    account = value,
-                    active = true,
-                )
-            })
-        }
-    }
+    suspend fun delete(entity: CloudEntity) = cloudDao.delete(entity)
 
-    suspend fun delete(entity: CloudEntity) = Rclone.Config.delete(entity.name).also { result ->
-        if (result.isSuccess) {
-            cloudDao.deleteCloud(entity)
-        }
-    }
-
-    fun getTmpMountPath(name: String) = pathUtil.getTmpMountPath(name)
-
-    suspend fun mountTmp(name: String, tmpMountPath: String) = run {
-        rootService.mkdirs(tmpMountPath)
-        Rclone.mount(
-            src = "${name}:",
-            dst = tmpMountPath,
-            "--allow-non-empty",
-            "--allow-other",
-            "--vfs-cache-mode off",
-            "--read-only",
-        )
-    }
-
-    suspend fun unmountTmp(tmpMountPath: String) = run {
-        BaseUtil.kill("rclone")
-        BaseUtil.umount(tmpMountPath)
-        rootService.deleteRecursively(tmpMountPath)
-    }
-
-    suspend fun upload(src: String, dstDir: String): ShellResult = run {
+    suspend fun upload(client: CloudClient, src: String, dstDir: String): ShellResult = run {
         log { "Uploading..." }
 
-        var isSuccess: Boolean
+        var isSuccess = true
         val out = mutableListOf<String>()
 
-        Rclone.copy(src = src, dst = dstDir).also { result ->
-            isSuccess = result.isSuccess
-            out.addAll(result.out)
+        runCatching {
+            client.upload(src = src, dst = dstDir)
+            out.add("Upload succeed.")
+        }.onFailure {
+            isSuccess = false
+            if (it.localizedMessage != null)
+                out.add(log { it.localizedMessage!! })
         }
+
         rootService.deleteRecursively(src).also { result ->
             isSuccess = isSuccess and result
             if (result.not()) out.add(log { "Failed to delete $src." })
@@ -100,27 +54,42 @@ class CloudRepository @Inject constructor(
         ShellResult(code = if (isSuccess) 0 else -1, input = listOf(), out = out)
     }
 
-    suspend fun download(src: String, dstDir: String, onDownloaded: suspend () -> Unit): ShellResult = run {
-        log { "Downloading..." }
+    suspend fun download(
+        client: CloudClient,
+        src: String,
+        dstDir: String,
+        onDownloaded: suspend () -> Unit,
+        deleteAfterDownloaded: Boolean = true,
+    ): ShellResult =
+        run {
+            log { "Downloading..." }
 
-        var code: Int
-        val out = mutableListOf<String>()
+            var code = 0
+            val out = mutableListOf<String>()
+            rootService.deleteRecursively(dstDir)
+            rootService.mkdirs(dstDir)
+            PathUtil.setFilesDirSELinux(context)
 
-        Rclone.copy(src = src, dst = dstDir).also { result ->
-            code = if (result.isSuccess) 0 else -2
-            out.addAll(result.out)
-        }
+            runCatching {
+                client.download(src = src, dst = dstDir)
+                out.add("Download succeed.")
+            }.onFailure {
+                code = -2
+                if (it.localizedMessage != null)
+                    out.add(log { it.localizedMessage!! })
+            }
 
-        if (code == 0) {
-            onDownloaded()
-        } else {
-            out.add(log { "Failed to download $src." })
-        }
-        rootService.deleteRecursively(dstDir).also { result ->
-            code = if (result) code else -1
-            if (result.not()) out.add(log { "Failed to delete $dstDir." })
-        }
+            if (code == 0) {
+                onDownloaded()
+            } else {
+                out.add(log { "Failed to download $src." })
+            }
+            if (deleteAfterDownloaded)
+                rootService.deleteRecursively(dstDir).also { result ->
+                    code = if (result) code else -1
+                    if (result.not()) out.add(log { "Failed to delete $dstDir." })
+                }
 
-        ShellResult(code = code, input = listOf(), out = out)
+            ShellResult(code = code, input = listOf(), out = out)
     }
 }
